Move utilities file to shared folder
