+++ src/bikespace_data/utilities/__init__.py
@@ -1,0 +1,1 @@
+from .utilities import StatusManager
